Add first draft of new review trigger
The new review trigger is slightly problematic because it triggers
on every review coming in. This might be nice for some people - but
for most people I imagine they just want to be notified that they
have reviews. I'll need to add a flag or some setting people can
use to set which style they prefer.

Also cleans up some random pieces around the codebase that aren't
needed.