Fix de-duplication on new review trigger
So turns out that making the id the first and last review id
doesn't actually solve the de-duplication problem. The thing is
we want to only notify users when they have new reviews once. As
they do reviews, if I used the ids of the reviews, it's possible
they would get notified as they reviewed said items. Now, however,
I query for only the reviews that are just now becoming available.
I then query again for all available reviews to get the total
numbers. Hopefully this solves the de-duplication issues and is
useful and makes sense butttt I might be back in an hour or in a
few days with everything changed again..